requirements.txt
python-jobspy>=1.1.80
anthropic>=0.45.0
requests>=2.31.0
weasyprint>=60.0
jinja2>=3.1.0
